+++ Adj/Program/Datastructure/Implementation.hs
@@ -1,3 +1,4 @@
 module Adj.Program.Datastructure.Implementation (module Exports) where
 
 import Adj.Program.Datastructure.Implementation.List as Exports
+import Adj.Program.Datastructure.Implementation.Tree as Exports