+++ Adj/Program/Datastructure/Implementation/Tree.hs
@@ -1,0 +1,7 @@
+module Adj.Program.Datastructure.Implementation.Tree where
+
+import Adj.Auxiliary (type (=!??=))
+import Adj.Program.Primitive.Maybe (Maybe)
+import Adj.Program.Primitive.Generation (Construction)
+
+type Tree = Maybe =!??= Construction