+++ CHANGELOG.md
@@ -204,3 +204,4 @@
 * Define `=:*:=` type alias
 * Define `=:+:=` type alias
 * Define `FGG` scheme
+* Define `Adj.Program.Datastructure.Implementation.Tree` module